# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RAGAcademic",
    "description": "Multimodal RAG chatbot untuk materi kuliah",
    "version": "0.3.0"
  },
  "paths": {
...
    "/catalog/courses": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Courses",
        "description": "Step 1 — daftar mata kuliah yang punya materi terindex.",
        "operationId": "get_courses_catalog_courses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/catalog/courses/{course_id}/weeks": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Weeks",
        "description": "Step 2 — minggu yang tersedia untuk satu mata kuliah (auto-update).",
        "operationId": "get_weeks_catalog_courses__course_id__weeks_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeekListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/courses/{course_id}/weeks/{week}/materials": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Materials",
        "description": "Step 3 — daftar materi (dokumen) untuk satu minggu.",
        "operationId": "get_materials_catalog_courses__course_id__weeks__week__materials_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "week",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Week"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/materials/{content_id}/{source_file}/starter-questions": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Starter Questions",
        "description": "Step 4 — pertanyaan template untuk satu materi (auto-generate + cache).\n\n`model` (query param) opsional — key model dari /models untuk memilih LLM.",
        "operationId": "get_starter_questions_catalog_materials__content_id___source_file__starter_questions_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          },
          {
            "name": "source_file",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source File"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StarterQuestionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/materials/{content_id}/{source_file}/quiz": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Quiz",
        "description": "Kuis pilihan ganda untuk satu materi (auto-generate + cache).\n\n`model` (query param) opsional. Tiap soal: pertanyaan, 4 opsi, answer_index, penjelasan.",
        "operationId": "get_quiz_catalog_materials__content_id___source_file__quiz_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          },
          {
            "name": "source_file",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source File"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/materials/{content_id}/{source_file}/quiz/submit": {
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Submit Quiz",
        "description": "Nilai jawaban kuis mahasiswa → skor + pembahasan per soal, dan simpan progres.\n\n`answers` = indeks opsi (0-3) yang dipilih, urut sesuai soal dari GET .../quiz.",
        "operationId": "submit_quiz_catalog_materials__content_id___source_file__quiz_submit_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Id"
            }
          },
          {
            "name": "source_file",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source File"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/chat/ask": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Ask Question",
        "operationId": "ask_question_chat_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/chat/feedback": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Submit Feedback",
        "operationId": "submit_feedback_chat_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseInfo": {
        "properties": {
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "course_name"
        ],
        "title": "CourseInfo"
      },
      "CourseListResponse": {
        "properties": {
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseInfo"
            },
            "type": "array",
            "title": "Courses"
          }
        },
        "type": "object",
        "required": [
          "courses"
        ],
        "title": "CourseListResponse"
      },
      "FeedbackRequest": {
        "properties": {
          "interaction_id": {
            "type": "string",
            "title": "Interaction Id"
          },
          "rating": {
            "type": "string",
            "enum": [
              "membantu",
              "cukup",
              "tidak_membantu"
            ],
            "title": "Rating"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "interaction_id",
          "rating"
        ],
        "title": "FeedbackRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaterialInfo": {
        "properties": {
          "source_file": {
            "type": "string",
            "title": "Source File"
          },
          "content_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Id"
          }
        },
        "type": "object",
        "required": [
          "source_file"
        ],
        "title": "MaterialInfo"
      },
      "MaterialListResponse": {
        "properties": {
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "week": {
            "type": "integer",
            "title": "Week"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/MaterialInfo"
            },
            "type": "array",
            "title": "Materials"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "week",
          "materials"
        ],
        "title": "MaterialListResponse"
      },
...
      "QueryRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "content_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Id"
          },
          "source_filter": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Filter"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "sederhana",
                  "standar",
                  "detail"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceInfo"
            },
            "type": "array",
            "title": "Sources"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "interaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interaction Id"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "sources",
          "recommendations",
          "session_id"
        ],
        "title": "QueryResponse"
      },
      "QuizQuestion": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "answer_index": {
            "type": "integer",
            "title": "Answer Index"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "question",
          "options",
          "answer_index"
        ],
        "title": "QuizQuestion"
      },
      "QuizResponse": {
        "properties": {
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "source_file": {
            "type": "string",
            "title": "Source File"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestion"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "content_id",
          "source_file",
          "questions"
        ],
        "title": "QuizResponse"
      },
      "QuizResultItem": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "your_answer": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Your Answer"
          },
          "correct_answer": {
            "type": "integer",
            "title": "Correct Answer"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "question",
          "options",
          "correct_answer",
          "is_correct"
        ],
        "title": "QuizResultItem"
      },
      "QuizSubmitRequest": {
        "properties": {
          "answers": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Answers",
            "description": "Indeks opsi yang dipilih per soal (urut sesuai soal)"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "student_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "QuizSubmitRequest"
      },
      "QuizSubmitResponse": {
        "properties": {
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "source_file": {
            "type": "string",
            "title": "Source File"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "attempt_id": {
            "type": "string",
            "title": "Attempt Id"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/QuizResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "content_id",
          "source_file",
          "total",
          "correct",
          "score",
          "attempt_id",
          "results"
        ],
        "title": "QuizSubmitResponse"
      },
      "SourceInfo": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "source_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source File"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "element_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Element Type"
          },
          "content_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Id"
          },
          "rerank_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Score"
          }
        },
        "type": "object",
        "required": [
          "index"
        ],
        "title": "SourceInfo"
      },
      "StarterQuestionsResponse": {
        "properties": {
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "source_file": {
            "type": "string",
            "title": "Source File"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "content_id",
          "source_file",
          "questions"
        ],
        "title": "StarterQuestionsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeekListResponse": {
        "properties": {
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "weeks": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Weeks"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "weeks"
        ],
        "title": "WeekListResponse"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}